Rewrite of tags system
Jesus christ... that was intense. 

Tags system has been rewritten to be able to run functions with parameter input. A few important notices:

Nesting may not function as intended, and generally some errors are to be expected since only so much can be done with regexing a string into arrays of various data types.

Also, only one tag per message is currently possible. See above notice. 

Tags are now also saved in a js file instead of a json file. And more...